+++ server/src/index.ts
@@ -22,6 +22,8 @@ import path from "path";
 import { adminRoutes } from "./admin/admin-auth.js";
 import PQueue from "p-queue";
 import { userRoutes } from "./routes/user-auth.js";
+import { fetchWorkshopCoverImage } from "./workshop/workshop-routes.js"
+
 import {
   deleteUserHistory,
   fetchPublicUserDatapack,
@@ -313,7 +315,7 @@ server.get<{ Params: { workshopId: number } }>(
       }
     }
   },
-  routes.fetchWorkshopCoverImage
+  fetchWorkshopCoverImage
 );
 
 server.register(adminRoutes, { prefix: "/admin" });
